Switch WAXPBY to take double scalars

+++ hpcg/src-mixed/ComputeWAXPBY.cpp
@@ -39,8 +39,8 @@
 
   @see ComputeWAXPBY_ref
 */
-int ComputeWAXPBY(const local_int_t n, const float alpha, const Vector & x,
-    const float beta, const Vector & y, Vector & w, bool & isOptimized) {
+int ComputeWAXPBY(const local_int_t n, const double alpha, const Vector & x,
+    const double beta, const Vector & y, Vector & w, bool & isOptimized) {
 
   // This line and the next two lines should be removed and your version of ComputeWAXPBY should be used.
   isOptimized = false;
