+++ hpcg/src-mixed/ComputeWAXPBY.hpp
@@ -15,6 +15,6 @@
 #ifndef COMPUTEWAXPBY_HPP
 #define COMPUTEWAXPBY_HPP
 #include "Vector.hpp"
-int ComputeWAXPBY(const local_int_t n, const float alpha, const Vector & x,
-    const float beta, const Vector & y, Vector & w, bool & isOptimized);
+int ComputeWAXPBY(const local_int_t n, const double alpha, const Vector & x,
+    const double beta, const Vector & y, Vector & w, bool & isOptimized);
 #endif // COMPUTEWAXPBY_HPP
