+++ hpcg/src-mixed/ComputeWAXPBY_ref.hpp
@@ -15,6 +15,6 @@
 #ifndef COMPUTEWAXPBY_REF_HPP
 #define COMPUTEWAXPBY_REF_HPP
 #include "Vector.hpp"
-int ComputeWAXPBY_ref(const local_int_t n, const float alpha, const Vector & x,
-    const float beta, const Vector & y, Vector & w);
+int ComputeWAXPBY_ref(const local_int_t n, const double alpha, const Vector & x,
+    const double beta, const Vector & y, Vector & w);
 #endif // COMPUTEWAXPBY_REF_HPP
